@@ -4,7 +4,7 @@
 
     <groupId>com.att.research.xacml</groupId>
     <artifactId>att-xacml</artifactId>
-    <version>3.2.1-SNAPSHOT</version>
+    <version>4.0.0-SNAPSHOT</version>
 
     <packaging>pom</packaging>
     <name>ATT XACML</name>
@@ -60,9 +60,9 @@
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
         
         <!-- Sonar doesn't seem to detect source level yet for java 11 -->
-        <maven.compiler.source>11</maven.compiler.source>
-        <maven.compiler.target>11</maven.compiler.target>
-        <maven.compiler.release>11</maven.compiler.release>
+        <maven.compiler.source>17</maven.compiler.source>
+        <maven.compiler.target>17</maven.compiler.target>
+        <maven.compiler.release>17</maven.compiler.release>
 
         <sonar-jacoco-listeners.version>3.2</sonar-jacoco-listeners.version>
         <sonar.core.codeCoveragePlugin>jacoco</sonar.core.codeCoveragePlugin>
@@ -72,7 +72,7 @@
         <version.gson>2.10.1</version.gson>
         <version.lombok>1.18.28</version.lombok>
         <version.httpclient>4.5.14</version.httpclient>
-        <version.servletapi>4.0.1</version.servletapi>
+        <version.servletapi>6.0.0</version.servletapi>
         <version.slf4j>1.7.30</version.slf4j>
         <version.logback>1.4.11</version.logback>
         <version.commons.io>2.13.0</version.commons.io>
@@ -253,12 +253,12 @@
                 <artifactId>lombok</artifactId>
                 <version>${version.lombok}</version>
             </dependency>
-            <dependency>
-                <groupId>javax.servlet</groupId>
-                <artifactId>javax.servlet-api</artifactId>
-                <version>${version.servletapi}</version>
-                <scope>provided</scope>
-            </dependency>
+			<dependency>
+			    <groupId>jakarta.servlet</groupId>
+			    <artifactId>jakarta.servlet-api</artifactId>
+			    <version>${version.servletapi}</version>
+			    <scope>provided</scope>
+			</dependency>
             <dependency>
                 <groupId>commons-io</groupId>
                 <artifactId>commons-io</artifactId>
@@ -333,7 +333,7 @@
                     <artifactId>maven-compiler-plugin</artifactId>
                     <version>3.11.0</version>
                     <configuration>
-                        <release>11</release>
+                        <release>17</release>
                     </configuration>
                 </plugin>
                 <plugin>
