@@ -5,5 +5,5 @@
 // Generated on: 2013.05.23 at 02:47:00 PM EDT 
 //
 
-@javax.xml.bind.annotation.XmlSchema(namespace = "urn:oasis:names:tc:xacml:3.0:core:schema:wd-17", elementFormDefault = javax.xml.bind.annotation.XmlNsForm.QUALIFIED)
+@jakarta.xml.bind.annotation.XmlSchema(namespace = "urn:oasis:names:tc:xacml:3.0:core:schema:wd-17", elementFormDefault = jakarta.xml.bind.annotation.XmlNsForm.QUALIFIED)
 package oasis.names.tc.xacml._3_0.core.schema.wd_17;
